Auth

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require("express");
 const multer = require("multer");
 
 const Post = require("../models/post");
+const checkAuth = require("../middleware/check-auth");
 
 const router = express.Router();
 
@@ -32,13 +33,15 @@ const storage = multer.diskStorage({
 
 router.post(
   "",
+  checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     const url = req.protocol + "://" + req.get("host");
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -48,28 +51,45 @@ router.post(
           id: createdPost._id
         }
       });
-    });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: 'Creating post failed!'
+      })
+    })
   }
 );
 
 router.put(
   "/:id",
+  checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
+
     if (req.file) {
       const url = req.protocol + "://" + req.get("host");
       imagePath = url + "/images/" + req.file.filename;
     }
+
     const post = new Post({
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.modifiedCount > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      }
+      res.status(401).json({ message: "Not authorized!" });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Cloudn't update post!"
+      })
     });
   }
 );
@@ -82,8 +102,7 @@ router.get("", (req, res, next) => {
   if (pageSize && currentPage) {
     postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
   }
-  postQuery
-    .then(documents => {
+  postQuery.then(documents => {
       fetchedPosts = documents;
       return Post.count();
     })
@@ -93,23 +112,42 @@ router.get("", (req, res, next) => {
         posts: fetchedPosts,
         maxPosts: count
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching posts failed"
+      })
     });
 });
 
 router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
+  Post.findById(req.params.id)
+  .then(post => {
     if (post) {
       res.status(200).json(post);
     } else {
       res.status(404).json({ message: "Post not found!" });
     }
-  });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching posts failed"
+    })
+  });;
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+router.delete("/:id", checkAuth,(req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.deletedCount > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    }
+    res.status(401).json({ message: "Not authorized!" });
+
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching posts failed"
+    })
   });
 });
